# openapi.yaml
openapi: 3.0.3
info:
  title: Book Worm E-Bookstore API
  description: |
    Complete API specification for the Book Worm E-Bookstore platform.
    This API provides endpoints for user authentication, book browsing, cart management, 
    wishlist, order processing, payments, reviews, coupons, and gift points.
  version: 1.0.0
  contact:
    name: Book Worm Support
    email: support@bookworm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.bookworm.com/v1
    description: Production server
  - url: https://staging-api.bookworm.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
paths:
...
  /books:
    get:
      tags: [Books]
      summary: Get all books
      operationId: getAllBooks
      parameters:
        - {name: category, in: query, schema: {type: string}}
        - {name: format, in: query, schema: {type: string, enum: [Paperback, Hardcover, eBook]}}
        - {name: language, in: query, schema: {type: string}}
        - {name: minPrice, in: query, schema: {type: number, minimum: 0}}
        - {name: maxPrice, in: query, schema: {type: number, minimum: 0}}
        - {name: sortBy, in: query, schema: {type: string, enum: [title, price, rating, author], default: title}}
        - {name: section, in: query, schema: {type: string, enum: [recommended, bestseller, newlaunch, trending]}}
        - {name: search, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: Books retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/Book'}}
                  pagination: {$ref: '#/components/schemas/Pagination'}
        '400': {$ref: '#/components/responses/BadRequest'}
  /books/{id}:
    get:
      tags: [Books]
      summary: Get single book by ID
      operationId: getBookById
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Book'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /categories/{id}/books:
    get:
      tags: [Categories]
      summary: Get books by category
      operationId: getBooksByCategory
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: sortBy, in: query, schema: {type: string, enum: [title, price, rating], default: title}}
      responses:
        '200':
          description: Books retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/Book'}}
                  pagination: {$ref: '#/components/schemas/Pagination'}
        '404': {$ref: '#/components/responses/NotFound'}

  # ==================== CART ====================
...
  /wishlist:
    get:
      tags: [Wishlist]
      summary: Get user's wishlist
      operationId: getWishlist
      security: [{bearerAuth: []}]
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
      responses:
        '200':
          description: Wishlist retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/Book'}}
                  pagination: {$ref: '#/components/schemas/Pagination'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
    post:
      tags: [Wishlist]
      summary: Add book to wishlist
      operationId: addToWishlist
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId]
              properties:
                bookId: {type: string}
      responses:
        '201':
          description: Book added to wishlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  message: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /orders:
    get:
      tags: [Orders]
      summary: Get user's order history
      operationId: getOrders
      security: [{bearerAuth: []}]
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned]}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 10}}
      responses:
        '200':
          description: Orders retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/Order'}}
                  pagination: {$ref: '#/components/schemas/Pagination'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
    post:
      tags: [Orders]
      summary: Place new order
      operationId: createOrder
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [addressId]
              properties:
                addressId: {type: string}
                couponCode: {type: string}
                giftPointsToRedeem: {type: integer, minimum: 0}
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Order'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
...
  /addresses:
    get:
      tags: [Addresses]
      summary: Get user's saved addresses
      operationId: getAddresses
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Addresses retrieved
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Address'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
    post:
      tags: [Addresses]
      summary: Add new address
      operationId: addAddress
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddressInput'}
      responses:
        '201':
          description: Address added
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Address'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
  /addresses/{id}:
    put:
      tags: [Addresses]
      summary: Update address
      operationId: updateAddress
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddressInput'}
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Address'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Addresses]
      summary: Delete address
      operationId: deleteAddress
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Address deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  message: {type: string}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Address:
      type: object
      required: [id, name, phone, street, city, state, zipCode, country]
      properties:
        id: {type: string, example: "addr_123"}
        name: {type: string, example: "John Doe"}
        phone: {type: string, example: "+1234567890"}
        street: {type: string, example: "123 Main St, Apt 4B"}
        city: {type: string, example: "New York"}
        state: {type: string, example: "NY"}
        zipCode: {type: string, example: "10001"}
        country: {type: string, example: "USA"}
        isDefault: {type: boolean, example: true}
        addressType: {type: string, enum: [home, office, other], example: "home"}
    AddressInput:
      type: object
      required: [name, phone, street, city, state, zipCode, country]
      properties:
        name: {type: string}
        phone: {type: string}
        street: {type: string}
        city: {type: string}
        state: {type: string}
        zipCode: {type: string}
        country: {type: string}
        addressType: {type: string, enum: [home, office, other]}
    Book:
      type: object
      required: [id, title, author, price, coverImage, description, category, format, rating, reviewCount]
      properties:
        id: {type: string, example: "book_123"}
        title: {type: string, example: "The Art of Focus"}
        author: {type: string, example: "Arjun Patel"}
        price: {type: number, format: float, example: 399.00}
        originalPrice: {type: number, format: float, example: 499.00}
        discount: {type: number, format: float, example: 20}
        coverImage: {type: string, format: uri}
        description: {type: string}
        category: {$ref: '#/components/schemas/Category'}
        genre: {$ref: '#/components/schemas/BookGenre'}
        format: {type: string, enum: [Paperback, Hardcover, eBook, Audiobook]}
        language: {type: string, example: "English"}
        rating: {type: number, format: float, minimum: 0, maximum: 5, example: 4.5}
        reviewCount: {type: integer, example: 234}
        pages: {type: integer, example: 320}
        publisher: {type: string, example: "Penguin Books"}
        publishDate: {type: string, format: date}
        isbn: {type: string, example: "978-3-16-148410-0"}
        tags: {type: array, items: {type: string}}
        inStock: {type: boolean, example: true}
        stockQuantity: {type: integer, example: 50}
        deliveryDate: {type: string, example: "Mon, 23 Jul"}
        isBestseller: {type: boolean}
        isNewLaunch: {type: boolean}
        isTrending: {type: boolean}
    BookGenre:
      type: object
      required: [id, name]
      properties:
        id: {type: string, example: "genre_123"}
        name: {type: string, example: "Self-help"}
        description: {type: string}
    Category:
      type: object
      required: [id, name]
      properties:
        id: {type: string, example: "cat_123"}
        name: {type: string, example: "Self-help"}
        description: {type: string}
        bookCount: {type: integer, example: 42}
        image: {type: string, format: uri}
...
    Order:
      type: object
      required: [id, userId, items, status, total, createdAt]
      properties:
        id: {type: string, example: "order_123"}
        userId: {type: string, example: "user_123"}
        orderNumber: {type: string, example: "ORD-2024-001"}
        items: {type: array, items: {$ref: '#/components/schemas/OrderItem'}}
        status: {type: string, enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned]}
        subtotal: {type: number, format: float}
        tax: {type: number, format: float}
        deliveryCharges: {type: number, format: float}
        discount: {type: number, format: float}
        total: {type: number, format: float}
        shippingAddress: {$ref: '#/components/schemas/Address'}
        paymentMethod: {type: string, enum: [credit_card, debit_card, upi, net_banking, cash_on_delivery, wallet]}
        paymentStatus: {type: string, enum: [pending, completed, failed, refunded]}
        trackingNumber: {type: string}
        estimatedDelivery: {type: string, format: date}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    OrderItem:
      type: object
      required: [id, book, quantity, price]
      properties:
        id: {type: string, example: "item_123"}
        book: {$ref: '#/components/schemas/Book'}
        quantity: {type: integer, minimum: 1}
        price: {type: number, format: float}
        subtotal: {type: number, format: float}
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string, example: "INVALID_REQUEST"}
        message: {type: string, example: "Invalid request parameters"}
        details: {type: object}
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: {type: integer, minimum: 1, example: 1}
        limit: {type: integer, minimum: 1, example: 20}
        total: {type: integer, minimum: 0, example: 100}
        totalPages: {type: integer, minimum: 0, example: 5}
        hasNext: {type: boolean, example: true}
        hasPrev: {type: boolean, example: false}
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
security:
  - bearerAuth: []